Renaming clear-cache to avoid confusion with clearing out the lancache's data
